requirements.txt
# FastAPI & server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# DB
sqlalchemy==2.0.36
aiosqlite==0.20.0

# RAG & embeddings
langchain>=0.3.9
langchain-community>=0.3.9
langchain-core>=0.3.21,<0.4
langchain-openai>=0.2.6
langchain-google-genai>=2.0.4
langchain-huggingface>=0.2.0,<0.3
chromadb>=0.5.0
sentence-transformers>=3.0.0

# OCR
pytesseract==0.3.13
Pillow==11.0.0
python-multipart==0.0.17

# Env & utils
pydantic==2.10.3
pydantic-settings==2.6.1
